Day 8 complete*
Cheated a bit on part 2, I didn't see how the math of the loops worked out, and I still don't.

It seems to me that if the cycles didn't all line up, this solution wouldn't work. I believe the instruction set is guaranteed to form a stable circle which passes over a Z location at the same frequency as the first time, which is not guaranteed for all possible inputs.